requirements.txt
numpy==1.24.2
pillow==9.5.0
tflite-runtime==2.11.0
